finalizing the project

+++ Controllers/DeleteController.php
@@ -1,6 +1,7 @@
 <?php
 require "../Bootstap/dbuser.php";
 $comming_id = intval($_REQUEST["id"]);
-$op = $dbuser->deleteUser($comming_id);
+if($dbOrder->deleteUserOrders($comming_id))
+    $op = $dbuser->deleteUser($comming_id);
 header("Location:../View/allUsers.php");
 exit;
